# pkg/models/member.go
package models

import "go.mongodb.org/mongo-driver/bson/primitive"

type Member struct {
	ID    primitive.ObjectID `bson:"_id,omitempty" json:"id,omitempty"`
	Name  string             `bson:"name" json:"name"`
	Email string             `bson:"email" json:"email"`
	Age   int                `bson:"age" json:"age"`
}